+++ src/lib/tables/Table.svelte
@@ -1,23 +1,18 @@
 <script lang="ts">
-  export let header: Array<string>;
-  export let divClass: string = 'relative overflow-x-auto shadow-md sm:rounded-lg';
-  export let tableClass: string = 'w-full text-sm text-left text-gray-500 dark:text-gray-400';
-  export let theadClass: string = 'text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400';
+	import classNames from 'classnames';
+	import { setContext } from 'svelte';
+	export let divClass: string = 'relative overflow-x-auto shadow-md sm:rounded-lg';
+	export let striped: boolean = false;
+	export let hoverable: boolean = false;
+	$: setContext('striped', striped);
+	$: setContext('hoverable', hoverable);
 </script>
 
 <div class={divClass}>
-  <table class={tableClass}>
-    <thead class={theadClass}>
-      <tr>
-        {#each header as column}
-          <th scope="col" class="px-6 py-3">
-            {column}
-          </th>
-        {/each}
-      </tr>
-    </thead>
-    <tbody>
-      <slot />
-    </tbody>
-  </table>
+	<table
+		{...$$restProps}
+		class={classNames('w-full text-left text-sm text-gray-500 dark:text-gray-400', $$props.class)}
+	>
+		<slot />
+	</table>
 </div>